RW-180 added SubjectCounter.
