+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/SubjectCounter.java
@@ -1,0 +1,99 @@
+package org.pmiops.workbench.cohortbuilder;
+
+import com.google.cloud.bigquery.QueryParameterValue;
+import com.google.cloud.bigquery.QueryRequest;
+import org.pmiops.workbench.cohortbuilder.querybuilder.FactoryKey;
+import org.pmiops.workbench.cohortbuilder.querybuilder.QueryParameters;
+import org.pmiops.workbench.config.WorkbenchConfig;
+import org.pmiops.workbench.model.SearchGroup;
+import org.pmiops.workbench.model.SearchGroupItem;
+import org.pmiops.workbench.model.SearchRequest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import javax.inject.Provider;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+@Service
+public class SubjectCounter {
+
+    @Autowired
+    private Provider<WorkbenchConfig> workbenchConfig;
+
+    private static final String COUNT_SQL_TEMPLATE =
+            "select count(distinct person_id) as count\n" +
+                    "  from `${projectId}.${dataSetId}.person` person\n" +
+                    " where 1 = 1\n";
+
+    private static final String UNION_TEMPLATE =
+            " union distinct\n";
+
+    private static final String INCLUDE_SQL_TEMPLATE =
+            "and person.person_id in (${includeSql})\n";
+
+    private static final String EXCLUDE_SQL_TEMPLATE =
+            "and not exists\n" +
+                    "(select 'x' from\n" +
+                    "(${excludeSql})\n";
+
+    public QueryRequest buildSubjectCounterQuery(SearchRequest request) {
+        Map<String, QueryParameterValue> params = new HashMap<>();
+        List<String> queryParts = new ArrayList<>();
+        String finalSql = COUNT_SQL_TEMPLATE;
+
+        // build query for included search groups
+        for (SearchGroup includeGroup : request.getIncludes()) {
+            finalSql = finalSql + INCLUDE_SQL_TEMPLATE;
+            for (SearchGroupItem includeItem : includeGroup.getItems()) {
+                QueryRequest queryRequest = QueryBuilderFactory
+                        .getQueryBuilder(FactoryKey.getType(includeItem.getType()))
+                        .buildQueryRequest(new QueryParameters()
+                                .type(includeItem.getType())
+                                .parameters(includeItem.getSearchParameters()));
+                params.putAll(queryRequest.getNamedParameters());
+                queryParts.add(queryRequest.getQuery());
+            }
+            finalSql = finalSql.replace("${includeSql}", String.join(UNION_TEMPLATE, queryParts));
+            queryParts = new ArrayList<>();
+        }
+
+        // build query for excluded search groups
+        for (SearchGroup excludeGroup : request.getExcludes()) {
+            finalSql = finalSql + EXCLUDE_SQL_TEMPLATE;
+            for (SearchGroupItem excludeItem : excludeGroup.getItems()) {
+                QueryRequest queryRequest = QueryBuilderFactory
+                        .getQueryBuilder(FactoryKey.getType(excludeItem.getType()))
+                        .buildQueryRequest(new QueryParameters()
+                                .type(excludeItem.getType())
+                                .parameters(excludeItem.getSearchParameters()));
+                params.putAll(queryRequest.getNamedParameters());
+                queryParts.add(queryRequest.getQuery());
+            }
+            finalSql = finalSql.replace("${excludeSql}", String.join(UNION_TEMPLATE, queryParts));
+            queryParts = new ArrayList<>();
+        }
+
+        return QueryRequest
+                        .newBuilder(filterBigQueryConfig(finalSql))
+                        .setNamedParameters(params)
+                        .setUseLegacySql(false)
+                        .build();
+    }
+
+    protected String filterBigQueryConfig(String sqlStatement, String tableName) {
+        String returnSql = sqlStatement.replace("${projectId}", workbenchConfig.get().bigquery.projectId);
+        returnSql = returnSql.replace("${dataSetId}", workbenchConfig.get().bigquery.dataSetId);
+        if (tableName != null) {
+            returnSql = returnSql.replace("${tableName}", tableName);
+        }
+        return returnSql;
+    }
+
+    protected String filterBigQueryConfig(String sqlStatement) {
+        return filterBigQueryConfig(sqlStatement, null);
+
+    }
+}
