+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/querybuilder/FactoryKey.java
@@ -15,14 +15,13 @@ public enum FactoryKey {
     CODES,
     /** TODO: this is temporary and will be removed when we figure out the conceptId mappings **/
     GROUP_CODES,
-    DEMO,
-    COUNT;
+    DEMO;
 
     private static final Map<String, Object> typeMap = Collections.unmodifiableMap(initializeMapping());
 
-    public static String getType(String type) {
+    public static FactoryKey getType(String type) {
         if (typeMap.containsKey(type)) {
-            return ((FactoryKey)typeMap.get(type)).name();
+            return ((FactoryKey)typeMap.get(type));
         }
         throw new IllegalArgumentException("Invalid type provided: " + type);
     }
@@ -43,7 +42,6 @@ public enum FactoryKey {
         tMap.put("ICD10", FactoryKey.CODES);
         tMap.put("CPT", FactoryKey.CODES);
         tMap.put("DEMO", FactoryKey.DEMO);
-        tMap.put("COUNT", FactoryKey.COUNT);
         return tMap;
     }
 }
